instagram.svg
This is an Instagram icon according to the Figma design. Originally had issues with the formatting of the code based on the design to match all other icons. I fixed the formatting of the icons on the Figma design to resolve this problem. In the original design, not all of the icons had drop shadows. Fixed it so that it is consistent with all icons.